Update Ramps n' stuff final.ipynb
Code written to simulate an RC car performing a hill climb on a ramp with a radius of 1.05m. The goal was to test different gear ratios as well as adding a different number of 3.75 kg masses to the rear of the car. A gearbox was then designed around these simulation results.